LayoutTests/imported/w3c/web-platform-tests/webrtc/RTCRtpReceiver-getParameters-expected.txt


PASS getParameters() with audio receiver
PASS getParameters() with video receiver
FAIL getParameters() with simulcast video receiver assert_greater_than: expected a number greater than 0 but got 0
